% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_modeling_and_fitting.R
\name{batch_modeling_and_fitting}
\alias{batch_modeling_and_fitting}
\title{Construct A Batch of Keras Models}
\usage{
batch_modeling_and_fitting(data_list, param_combination, ...)
}
\arguments{
\item{data_list}{A \code{list} containing predictor and label matrix of training data and test data.
Please use \link{prepare_data} to generate this.}

\item{param_combination}{A parameter \code{matrix}/\code{data.frame} with each row representing the parameters
for run Keras model once. Column names should indicate parameter names and should be same as in modeling function.
\code{\link[base:expand.grid]{base::expand.grid()}} may be very useful to generate it.}

\item{...}{Other arguments passing to \link{modeling_and_fitting}.}
}
\value{
a \code{tibble}.
}
\description{
Construct A Batch of Keras Models
}
\examples{
load(system.file("extdata", "wang2020-input.RData",
  package = "sigminer.prediction", mustWork = TRUE
))
dat_list <- prepare_data(expo_all,
  col_to_vars = c(paste0("Sig", 1:5), paste0("AbsSig", 1:5)),
  col_to_label = "enrich_sig",
  label_names = paste0("Sig", 1:5)
)
pc <- expand.grid(
  c(10, 20, 50, 100),
  c(0, 0.1, 0.2, 0.3, 0.4, 0.5),
  c(10, 20, 50, 100),
  c(0, 0.1, 0.2, 0.3, 0.4, 0.5)
)
colnames(pc) <- c(
  "first_layer_unit", "second_layer_drop_rate",
  "third_layer_unit", "fourth_layer_drop_rate"
)

# Just use 2 rows for illustration
batch_res <- batch_modeling_and_fitting(dat_list, param_combination = pc \%>\% head(2))
batch_res

tidy(batch_res)
}
