% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{copy_model}
\alias{copy_model}
\title{Copy Model File}
\usage{
copy_model(model_file, dest)
}
\arguments{
\item{model_file}{A file path to the model file.}

\item{dest}{The destination file path.}
}
\value{
Nothing
}
\description{
It is usefully when your result model file is stored in temp directory and
you want to keep it.
}
