% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_trained_models.R
\name{list_trained_models}
\alias{list_trained_models}
\title{List Current Available Trained Keras Models}
\usage{
list_trained_models()
}
\value{
A \code{tibble} containing summary models.
}
\description{
List Current Available Trained Keras Models
}
\examples{
list_trained_models()
}
