% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_trained_model.R
\name{load_trained_model}
\alias{load_trained_model}
\title{Load Trained Models}
\usage{
load_trained_model(x)
}
\arguments{
\item{x}{A subset from \link{list_trained_models}.}
}
\value{
A (list of) Keras model.
}
\description{
Load Trained Models
}
\examples{
z <- list_trained_models() \%>\%
  head(1) \%>\%
  load_trained_model()
z
}
