% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare Training and Test Dataset}
\usage{
prepare_data(
  data,
  col_to_vars,
  col_to_label,
  label_names,
  seed = 1234,
  test_split = 0.2
)
}
\arguments{
\item{data}{A \code{data.frame}.}

\item{col_to_vars}{A character vector specifying the predictive columns.}

\item{col_to_label}{A column indicating the labels/classes.}

\item{label_names}{Label/class names. The order is important. For example,
"a", "b", "c" will be transformed to 0, 1, 2.}

\item{seed}{Random seed, default is \code{1234}.}

\item{test_split}{A fraction of samples to treated as test dataset, default is \code{0.2}.}
}
\value{
a \code{list} containing \code{x_train}, \code{y_train}, \code{x_test}, \code{y_test} datasets.
}
\description{
Prepare Training and Test Dataset
}
\examples{
load(system.file("extdata", "wang2020-input.RData",
  package = "sigminer.prediction", mustWork = TRUE
))
dat_list <- prepare_data(expo_all,
  col_to_vars = c(paste0("Sig", 1:5), paste0("AbsSig", 1:5)),
  col_to_label = "enrich_sig",
  label_names = paste0("Sig", 1:5)
)
str(dat_list)
}
