% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy}
\alias{tidy}
\title{Tidy Modeling Result}
\usage{
tidy(x)
}
\arguments{
\item{x}{A result \code{tibble} from either \link{modeling_and_fitting} or \link{batch_modeling_and_fitting}.}
}
\value{
a \code{tibble}
}
\description{
Tidy Modeling Result
}
